fix(frontend): reverted change with dynamic components as it caused them to be rendered twice on page load
