+++ apps/frontend/src/blocks/PaymentsSummary/PaymentsSummary.server.tsx
@@ -1,9 +1,13 @@
+import dynamic from 'next/dynamic';
 import React from 'react';
 
 import { sdk } from '@/api/sdk';
 
-import { PaymentsSummaryDynamic } from './PaymentsSummary.dynamic';
 import { PaymentsSummaryProps } from './PaymentsSummary.types';
+
+export const PaymentsSummaryDynamic = dynamic(() =>
+    import('./PaymentsSummary.client').then((module) => module.PaymentsSummaryPure),
+);
 
 export const PaymentsSummary: React.FC<PaymentsSummaryProps> = async ({ id, accessToken, locale }) => {
     const data = await sdk.blocks.getPaymentsSummary(
