--- apps/frontend/src/blocks/TicketDetails/TicketDetails.dynamic.tsx
@@ -1,9 +1,0 @@
-'use client';
-
-import dynamic from 'next/dynamic';
-
-// an intermediary component is required for the dynamic import to work propertly with server components
-// @see https://github.com/vercel/next.js/issues/61066
-export const TicketDetailsDynamic = dynamic(() =>
-    import('./TicketDetails.client').then((module) => module.TicketDetailsPure),
-);
