+++ apps/frontend/src/blocks/TicketDetails/TicketDetails.server.tsx
@@ -1,9 +1,13 @@
+import dynamic from 'next/dynamic';
 import React from 'react';
 
 import { sdk } from '@/api/sdk';
 
-import { TicketDetailsDynamic } from './TicketDetails.dynamic';
 import { TicketDetailsProps } from './TicketDetails.types';
+
+export const TicketDetailsDynamic = dynamic(() =>
+    import('./TicketDetails.client').then((module) => module.TicketDetailsPure),
+);
 
 export const TicketDetails: React.FC<TicketDetailsProps> = async ({ id, ticketId, accessToken, locale }) => {
     const data = await sdk.blocks.getTicketDetails(
