+++ apps/frontend/src/blocks/TicketList/TicketList.server.tsx
@@ -1,9 +1,11 @@
+import dynamic from 'next/dynamic';
 import React from 'react';
 
 import { sdk } from '@/api/sdk';
 
-import { TicketListDynamic } from './TicketList.dynamic';
 import { TicketListProps } from './TicketList.types';
+
+export const TicketListDynamic = dynamic(() => import('./TicketList.client').then((module) => module.TicketListPure));
 
 export const TicketListServer: React.FC<TicketListProps> = async ({ id, accessToken, locale }) => {
     const data = await sdk.blocks.getTicketList(
