+++ apps/frontend/src/blocks/TicketRecent/TicketRecent.server.tsx
@@ -1,9 +1,15 @@
+import dynamic from 'next/dynamic';
 import React from 'react';
 
 import { sdk } from '@/api/sdk';
 
-import { TicketRecentPure } from './TicketRecent.client';
 import { TicketRecentProps } from './TicketRecent.types';
+
+// an intermediary component is required for the dynamic import to work propertly with server components
+// @see https://github.com/vercel/next.js/issues/61066
+export const TicketRecentDynamic = dynamic(() =>
+    import('./TicketRecent.client').then((module) => module.TicketRecentPure),
+);
 
 export const TicketRecent: React.FC<TicketRecentProps> = async ({ id, accessToken, locale }) => {
     const data = await sdk.blocks.getTicketRecent(
@@ -14,5 +20,5 @@ export const TicketRecent: React.FC<TicketRecentProps> = async ({ id, accessToke
         accessToken,
     );
 
-    return <TicketRecentPure {...data} id={id} accessToken={accessToken} locale={locale} />;
+    return <TicketRecentDynamic {...data} id={id} accessToken={accessToken} locale={locale} />;
 };
