+++ apps/frontend/turbo/generators/templates/block/server.hbs
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { sdk } from '@/api/sdk';
 
-import {  {{ pascalCase name }}Dynamic } from './{{ pascalCase name }}.dynamic';
+import {  {{ pascalCase name }}Pure } from './{{ pascalCase name }}.client';
 import {  {{ pascalCase name }}Props } from './{{ pascalCase name }}.types';
 
 export const {{ pascalCase name }}: React.FC<{{ pascalCase name }}Props> = async ({ id, accessToken, locale }) => {
